requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pymysql==1.1.1
sqlalchemy==2.0.36
python-dotenv==1.0.1
pydantic==2.10.3
python-multipart==0.0.18
